update on the prosilica test executable
The test executable now returns 0 on successfully opening the camera by uid (may be supplied as command line argument) and -1 otherwise.
This was done to enable the program to be used in a basic hardware test script.
Added an install statement for the prosilica test executable to the src/CMakeLists.txt